agents/replygenie.py
import os
from dotenv import load_dotenv

class ReplyGenie:
    """
    Drafts personalized responses for client queries, adapting to tone and context.
    """
    def __init__(self):
        load_dotenv()
        self.openai_key = os.getenv('OPENAI_API_KEY')

    def draft_response(self, client_query, context, tone='friendly'):
        # Placeholder: Use GPT-4 to draft a personalized response
        pass 